==> 11_observations.qmd <==
```{r setup}
#| message: false
#| include: false
library(tidyverse)
library(knitr)
library(googlesheets4)
```

# Observations {.unnumbered}

Data exploration.

```{r data}
#| message: false
#| warning: false
#| eval: false
metadata <- read_sheet("https://docs.google.com/spreadsheets/d/1fq2owxMBLBwwibcdw2uQQFxnIhsMbaH4Qcj_xUwVvSQ/edit?gid=0#gid=0", col_types = "c") %>% # nolint
  rename_all(tolower) %>%
  select(site, plot, treatment, year_of_harvest) %>%
  separate(year_of_harvest, "year_of_harvest") %>%
  mutate(year_of_harvest = as.numeric(year_of_harvest)) %>%
  rename(harvest_year = year_of_harvest) %>%
  mutate(treatment = ifelse(
    treatment %in% c("control", "Control", "Natural_Forest"),
    "control", "logged"
  )) %>%
  group_by(site) %>%
  mutate(harvest_year_min = min(harvest_year, na.rm = TRUE)) %>%
  mutate(harvest_year = ifelse(site == "Paracou", 1986, harvest_year)) %>%
  mutate(harvest_year_min = ifelse(site == "Paracou", 1986, harvest_year_min))
data <- read_csv("data/raw_data/aggregated_data.csv") %>%
  rename_all(tolower) %>%
  left_join(metadata) %>%
  filter(!(site %in% c(
    "Peteco", "Bafog", "Tene", "Sg Lalang",
    "Montagne_Tortue", "Manare"
  ))) %>%
  filter(!(site == "Kabo" & year < 1980)) %>%
  mutate(rel_year = ifelse(treatment == "control",
    year - harvest_year_min,
    year - harvest_year
  )) %>%
  mutate(sitenum = as.numeric(as.factor(site))) %>%
  group_by(site) %>%
  mutate(y = ifelse(variable == "agb" & site == "Kibale",
    300, value
  )) %>%
  ungroup()
data_ogf <- data %>%
  filter((rel_year <= 0 & treatment == "logged") | treatment == "control") %>%
  mutate(plotnum = as.numeric(as.factor(paste(site, plot))))
data_rec <- data %>%
  filter(treatment == "logged", rel_year > 2) %>%
  mutate(plotnum = as.numeric(as.factor(paste0(site, "_", plot))))
write_tsv(data, "data/derived_data/data.tsv")
write_tsv(data_rec, "data/derived_data/data_rec.tsv")
write_tsv(data_ogf, "data/derived_data/data_ogf.tsv")
```

Computed variables are:

```{r vars}
#| output: asis
read_tsv("data/derived_data/data.tsv", col_types = cols()) %>%
  select(variable) %>%
  unique() %>%
  unlist() %>%
  paste0(collapse = ", ")
```

```{r fig1}
#| message: false
#| warning: false
#| fig-cap: "All modeled attributes trajectories."
vars <- c( # nolint
  "nstem", "agb", "ba", "gini", # nolint
  "diversity_q1_gen", "FDiv", "FEve", # nolint
  "WD_cwm_ba", "SLA_cwm_ba", "N_cwm_ba" # nolint
)
read_tsv("data/derived_data/data.tsv") %>%
  filter(variable %in% vars) %>%
  ggplot(aes(year, value, group = paste(site, plot))) +
  geom_line() +
  theme_bw() +
  facet_wrap(~variable, scales = "free_y") +
  theme(
    axis.title = element_blank(),
    strip.text = element_text(size = 6)
  )
```

```{r figs_out}
#| eval: false
make_fig <- function(var) {
  g <- read_tsv("data/derived_data/data.tsv") %>%
    filter(variable == var) %>%
    ggplot(aes(year, value, group = paste(site, plot), col = treatment)) +
    geom_line() +
    theme_bw() +
    xlab("") +
    ylab("") +
    facet_wrap(~site, scales = "free") +
    geom_vline(aes(xintercept = harvest_year)) +
    theme(legend.position = "bottom") +
    ggtitle(var)
  ggsave(plot = g, filename = paste0("figures/obs/", var, ".png"))
}
unique(read_tsv("data/derived_data/data.tsv")$variable) %>%
  lapply(make_fig)
```

```{r fig3}
#| message: false
#| warning: false
#| fig-cap: "AGB trajectories across sites and treatments."
read_tsv("data/derived_data/data.tsv") %>%
  filter(variable == "agb") %>%
  ggplot(aes(year, value, group = paste(site, plot), col = treatment)) +
  geom_line() +
  theme_bw() +
  xlab("") +
  ylab(expression("AGB [" ~ Mg ~ ha^{
    -1
  } ~ "]")) +
  facet_wrap(~site, scales = "free") +
  geom_vline(aes(xintercept = harvest_year)) +
  theme(legend.position = "bottom")
```
