==> 12_data_prep.qmd <==
```{r setup}
#| message: false
#| include: false
library(tidyverse)
library(knitr)
aggregated_file <- "data/raw_data/aggregated_data_v9.csv"
```

# Data prep {.unnumbered}

Data exploration.

```{r data}
#| message: false
#| warning: false
#| eval: false
control_names <- c("Control")
metadata <- read_tsv("data/raw_data/metadata.tsv") %>%
  rename(site_raw = site) %>%
  left_join(read_tsv("data/raw_data/sites.tsv")) %>%
  select(-site_raw) %>%
  separate(year_of_harvest, "year_of_harvest") %>%
  mutate(year_of_harvest = as.numeric(year_of_harvest)) %>%
  rename(harvest_year = year_of_harvest) %>%
  mutate(treatment = ifelse(treatment %in% control_names,
    "control", "logged"
  )) %>%
  group_by(site) %>%
  mutate(harvest_year_min = min(harvest_year, na.rm = TRUE)) %>%
  mutate(harvest_year = ifelse(site == "Paracou", 1986, harvest_year)) %>%
  mutate(harvest_year_min = ifelse(site == "Paracou", 1986, harvest_year_min))
vars <- c(
  "ba", "agb", "nstem", "gini",
  "diversity_q1_gen", "diversity_q1_sp", "diversity_q0_gen", "diversity_q0_sp",
  "FDiv",
  "WD_cwm_ba", "SLA_cwm_ba"
)
sites <- c(
  "Antimary APU2", "Antimary APU3", "BAFOG", "Corinto", "Ecosilva", "Gola",
  "Iracema", "Iwokrama", "Jari", "Jenaro Herrera", "Kabo", "Kibale", "Lesong",
  "Malinau", "Manare", "Mbaiki", "Mil", "Misiones", "Moju", "Montagne Tortue",
  "Nelliyampathy", "Pad Limo 2 Barracos", "Pad Limo Chico Bocao",
  "Pad Limo Cumaru",
  "Pad Limo Jatoba", "Pad Limo Pocao", "Pad Limo STCP", "Paracou", "Peixoto",
  "Peteco", "Sao Nicolau", "Sungai Lalang", "Tapajos Km114",
  "Tapajos Km67", "Tirimbina", "Ulu Muda", "Uppangala", "SUAS"
)
antimary_sites <- c(
  "Antimary APU2", "Antimary APU3"
)
pad_sites <- c(
  "Pad Limo 2 Barracos", "Pad Limo Chico Bocao", "Pad Limo Cumaru",
  "Pad Limo Jatoba", "Pad Limo Pocao", "Pad Limo STCP"
)
spurious <- read_tsv("data/raw_data/spurious.tsv")
data <- read_csv(aggregated_file) %>%
  na.omit() %>%
  rename_all(tolower) %>%
  rename(site_raw = site) %>%
  left_join(read_tsv("data/raw_data/sites.tsv")) %>%
  select(-site_raw) %>% 
  rename(new_plot = plot) %>% 
  left_join(read_tsv("data/raw_data/plot_metadata.tsv")) %>% 
  left_join(metadata) %>% 
  select(-plot) %>% 
  rename(plot = new_plot, area = new_plot_area) %>% 
  filter(site %in% sites) %>%
  filter(!(site %in% spurious$site & 
             plot %in% spurious$plot & 
             year %in% spurious$year)) %>% 
  mutate(plot = ifelse(site %in% c(pad_sites, antimary_sites),
    paste(site, plot),
    plot
  )) %>%
  mutate(site = ifelse(site %in% pad_sites, "Pad Limo", site)) %>%
  mutate(site = ifelse(site %in% antimary_sites, "Antimary", site)) %>%
  filter(variable %in% vars) %>%
  mutate(rel_year = ifelse(treatment == "control",
    year - harvest_year_min,
    year - harvest_year
  )) %>%
  mutate(y = value) %>%
  mutate(sitenum = as.numeric(as.factor(site)))
data_equ <- data %>%
  filter((rel_year <= 0 & treatment == "logged") | treatment == "control") %>%
  group_by(variable) %>%
  mutate(plotnum = as.numeric(as.factor(paste(site, plot))))
data_rec <- data %>%
  filter(treatment == "logged", rel_year > 2) %>%
  group_by(variable) %>%
  mutate(plotnum = as.numeric(as.factor(paste0(site, "_", plot))))
write_tsv(data, "data/derived_data/data.tsv")
write_tsv(data_rec, "data/derived_data/data_rec.tsv")
write_tsv(data_equ, "data/derived_data/data_equ.tsv")
```

```{r figs}
#| message: false
#| warning: false
#| fig-cap: "Trajectories across sites and treatments."
#| fig-height: 10
#| fig-width: 10
vars <- c(
  "ba", "agb", "nstem", "gini",
  "diversity_q1_gen", "diversity_q1_sp", "diversity_q0_gen", "diversity_q0_sp",
  "FDiv",
  "WD_cwm_ba", "SLA_cwm_ba"
)
names(vars) <- vars
make_fig <- function(var) {
  read_tsv("data/derived_data/data.tsv") %>%
    filter(variable == var) %>%
    ggplot(aes(year, value, group = paste(site, plot), col = treatment)) +
    geom_line() +
    geom_point() +
    theme_bw() +
    xlab("") +
    ylab("") +
    facet_wrap(~site, scales = "free") +
    geom_vline(aes(xintercept = harvest_year)) +
    theme(legend.position = "bottom")
}
lapply(vars, make_fig)
```
