==> 21_models.qmd <==
```{r set}
#| message: false
#| include: false
library(tidyverse)
```

# Models {.unnumbered}

Models descriptions.

## Equilibirum

Equilibrium model using control and pre-logging inventories to inform $\theta$ prior of the recovery trajectory:

$$
\begin{align}
y \sim \mathcal N (\mu_{site}+\delta_{plot}, \sigma)\\
\delta_{plot} \sim \mathcal N(0,\sigma_{plot}) 
\end{align}
$$

## Recovery

Recovery trajectory model:

$$
\begin{align}
log(y) \sim \mathcal N (log(\mu), \sigma)\\
\mu = \theta_{site} \times (\phi_{plot} + LTP \times (1 - \phi_{plot}) + STP) \\
LTP = 1 - e^{-\lambda_{plot} \times t} \\
STP = \delta_{plot} \times (\frac{t}{\tau_{site}} \times e^{1-\frac{t}{\tau_{site}}})^2 \\
\lambda_{plot} \sim \mathcal N (\mu_{\lambda}, \sigma_{\lambda})\\
\tau_{site} \sim \mathcal N (\mu_{\tau}, \sigma_{\tau}) \\
(\sigma,\sigma_{\lambda},\sigma_{\tau}) \sim \mathcal N^3(0,1) \\
\theta_{site} \sim \mathcal N (mean(y_{control,site}), sd(y_{control,site})) \\
\phi_{plot} \sim \mathcal U(0.1,1.5) \\
\lambda_{plot} \sim \mathcal U(0,0.5) \\
\delta_{plot} \sim \mathcal U(-\phi_{plot},2) \\
\tau_{site} \sim \mathcal U(5,40) 
\end{align}
$$

The model we use is derived from the common framework for modelling recovery in disturbed tropical forests [@maurent2023]. The forest attribute is first disturbed at $t=0$ with either an increase or a decrease depending on the parameter $\phi$. Then the attributes recover asymptotically in the long term to an equilibrium value $\theta$ with a recovery rate $\lambda$. But the recovery trajectory can also accommodate a short term increase of decrease of the attribute at a time $\tau$ with an intensity $\delta$.

```{r recovery}
#| warning: false
#| message: false
#| fig-cap: "STP proj with pars."
ltp <- function(time, dist, thetainf, lambda) {
  theta0 <- thetainf * dist
  ltp <- 1 - exp(-lambda * time)
  theta0 + (thetainf - theta0) * (ltp)
}
stp <- function(time, dist, thetainf, lambda, tau, delta) {
  theta0 <- thetainf * dist
  ltp <- 1 - exp(-lambda * time)
  stp <- delta * (time / tau * exp(1 - time / tau))^2
  theta0 + ltp * (thetainf - theta0) + stp * thetainf
}
data.frame(time = seq(0.1, 50, length.out = 100)) %>%
  mutate(y_stp = stp(time,
    dist = 0.5, thetainf = 20, lambda = 0.05,
    tau = 5, delta = 0.5
  )) %>%
  mutate(y_ltp = ltp(time, dist = 0.5, thetainf = 20, lambda = 0.05)) %>%
  bind_rows(data.frame(time = 0, y_stp = 20, y_ltp = 20)) %>%
  ggplot(aes(time)) +
  geom_line(aes(y = y_ltp), col = "grey") +
  geom_line(aes(y = y_stp)) +
  theme_bw() +
  annotate("text",
    x = 52, y = 19.5, label = "theta[site]",
    col = "#9ea1d4", parse = TRUE, size = 5
  ) +
  annotate("text",
    x = 20, y = 15, label = "lambda[plot]",
    col = "#a8d2d1", parse = TRUE, size = 5
  ) +
  annotate("text",
    x = -4, y = 15, label = "phi[plot]",
    col = "#fc8a8a", parse = TRUE, size = 5
  ) +
  theme(axis.title = element_blank()) +
  geom_segment(aes(x = -2, y = 10, xend = -2, yend = 20),
    arrow = arrow(length = unit(0.2, "cm")), col = "#fc8a8a"
  ) +
  geom_segment(aes(x = -2, y = 20, xend = -2, yend = 10),
    arrow = arrow(length = unit(0.2, "cm")), col = "#fc8a8a"
  ) +
  annotate("text",
    x = 6, y = 9.3, label = "tau[site]",
    col = "#f2e5a9", parse = TRUE, size = 5
  ) +
  annotate("text",
    x = 4, y = 17.5, label = "delta[plot]",
    col = "#a7d6ac", parse = TRUE, size = 5
  ) +
  geom_segment(aes(x = 6, y = 13, xend = 6, yend = 22),
    arrow = arrow(length = unit(0.2, "cm")), col = "#a7d6ac"
  ) +
  geom_segment(aes(x = 6, y = 22, xend = 6, yend = 13),
    arrow = arrow(length = unit(0.2, "cm")), col = "#a7d6ac"
  ) +
  geom_segment(aes(x = 6, y = 9, xend = 6, yend = 7),
    arrow = arrow(length = unit(0.2, "cm")), col = "#f2e5a9"
  ) +
  theme(axis.title = element_blank()) +
  geom_segment(aes(x = -5, y = 20, xend = 0, yend = 20)) +
  ylim(7, NA)
```

> Choices (**to be developed as text in an appendix**):
>
> -   Using the log of y and a normal distribution centred on log(mu)
> -   Cauchy hyperdist
> -   Forced dist_p\<1
> -   Bounds of theta
> -   Free tau
> -   Free delta
>
> ⚠️ Hump is mathematically opposed to the sign of dist_p-1.
