==> 30_trajectories.qmd <==
```{r set}
#| message: false
#| include: false
library(tidyverse)
library(knitr)
library(cmdstanr)
list.files("r", full.names = TRUE, pattern = ".R") %>%
  lapply(source)
```

# Trajectories

This section develop trajectories per attributes.

```{r fit_all}
#| eval: false
equ <- cmdstan_model("models/equilibirum.stan")
rec_pos <- cmdstan_model("models/recovery_positive.stan")
rec <- cmdstan_model("models/recovery.stan")
vars <- c(
  "ba", "agb", "nstem", "gini",
  "diversity_q1_gen", "diversity_q1_sp", "diversity_q0_gen", "diversity_q0_sp", 
  "FDiv",
  "WD_cwm_ba", "SLA_cwm_ba"
)
vars <- c(
  "WD_cwm_ba"
)
lapply(vars, sample_equ)
lapply(vars, sample_rec)
```
