==> 51_wd.qmd <==
```{r set}
#| message: false
#| include: false
library(tidyverse)
library(knitr)
library(cmdstanr)
library(bayesplot)
library(loo)
library(patchwork)
list.files("r", full.names = TRUE, pattern = ".R") %>%
  lapply(source)
```

```{r}
#| eval: false
data$fit_rec$draws("tau0_s[21]") %>% 
  mcmc_trace() +
  scale_y_sqrt()

data$fit_rec$draws("tau0_s") %>% 
  mcmc_trace() +
  scale_y_sqrt()
```

```{r}
#| eval: false
data$fit_rec$summary("delta_s") %>% 
  separate(variable, c("delta", "s", "sitenum"), convert = TRUE) %>% 
  left_join(data$ind_rec) %>% 
  filter(site == 'Paracou') %>% 
  ggplot(aes(median)) + 
  geom_histogram() + 
  theme_bw() +
  facet_wrap(~ site, scales = "free_y") +
  theme(axis.text.x = element_text(angle = 90)) +
  scale_x_sqrt()
```

```{r}
#| eval: false
data$fit_rec$summary("lambda_p") %>% 
  separate(variable, c("delta", "p", "plotnum"), convert = TRUE) %>% 
  left_join(data$ind_rec) %>% 
  ggplot(aes(median)) + 
  geom_histogram() + 
  theme_bw() +
  facet_wrap(~ site, scales = "free_y") +
  theme(axis.text.x = element_text(angle = 90))
```

# Func - WD {.unnumbered}

Wood Density trajectories have been quantified using the community weighted mean, the average of the local distribution of a trait in a community. The trait value is weighted by tree basal area to better represent their biomass contribution in the community.

```{r data}
#| message: false
#| warning: false
var <- "WD_cwm_ba"
lab <- expression("WD [ g" ~ cm^{
  -3
} ~ "]")
data <- load_all(var, lab)
```

```{r theta_equ}
#| message: false
#| warning: false
#| fig-cap: "Theta posterior from forests at equilibirum."
equ_theta_fig(data)
```

```{r fit_rec}
#| message: false
#| warning: false
data$fit_rec
```

```{r pvo_stp}
#| message: false
#| warning: false
#| fig-cap: "Predicted versus observed."
pvo_fig(data)
```

```{r preds_stp}
#| message: false
#| warning: false
#| fig-cap: "Predicted trajectories."
#| fig-width: 10
#| fig-height: 10
proj_fig(data)
```

```{r theta_stp}
#| message: false
#| warning: false
#| fig-cap: "Theta posterior from recovery trajectories against its posterior."
rec_theta_fig(data)
```
