==> 61_group1.qmd <==
```{r set}
#| message: false
#| include: false
library(tidyverse)
library(knitr)
```

# Group 1 {.unnumbered}

Group 1 on **LoggingDiversity** focus on the **effects of logging disturbance on tropical forest diversity - a global assessment** with the following question: What is the effect of logging intensity on biodiversity, and how does it change through time? We started from two hypotheses:

-   **H1**: tree diversity long-term response increases through time to a maximum value, after which it then declines in late succession
-   **H2**: tree diversity also shows a non-linear response to increased intensity of logging: lightly logged forests show a transient increase in tree diversity, but more intensively logged forests show a strong negative effect on diversity

We will thus mainly work with diversity trajectories (genus & functional) bump characterized through bump time $\tau$ & bump intensity $\delta$ in the model and link it with forest structure (BA, AGB, or stem) disturbance index $\phi_{ba}$.

## H1 - Diversity & time

> **H1**: tree diversity long-term response increases through time to a maximum value, after which it then declines in late succession

**Kibale and Mbaiki have been removed has they don't have short term data. But, do we trust Malinau's bump with solely three recent inventories?** Genus diversity are good and could be trusted even if a few outlier sites delta might be underestimated, but is not a data issue? However, functional diversity trajectory is still not good enough needing either model or data tweaking. **Genus diversity delta distribution are bimodal and should be explored, e.g. I don't trust low delta in Paracou that are the sites lacking genus identification. Also how much plot delta max per site is linked to tau site being itself linked to the duration of inventories?**

```{r tau}
#| message: false
#| warning: false
read_tsv("outputs/parameters.tsv") %>%
  filter(level == "site") %>%
  filter(!(site %in% c("Kibale", "Uppangala"))) %>%
  filter(type == "diversity" & parameter == "tau") %>%
  ggplot(aes(x = site, col = region)) +
  geom_point(aes(y = median)) +
  geom_segment(aes(y = median - sd, yend = median + sd)) +
  geom_segment(aes(y = q5, yend = q95), size = .1) +
  facet_wrap(~attr_long) +
  coord_flip() +
  theme_bw() +
  ylab(expression(tau ~ "[ year ]")) +
  xlab("") +
  scale_color_discrete("") +
  ggtitle(expression(Site ~ tau)) +
  theme(legend.position = "bottom")
```

```{r delta}
#| message: false
#| warning: false
read_tsv("outputs/parameters.tsv") %>%
  filter(level == "plot") %>%
  filter(!(site %in% c("Kibale", "Uppangala"))) %>%
  filter(type == "diversity" & parameter == "deltapct") %>%
  ggplot(aes(x = median, y = site, fill = region)) +
  ggridges::geom_density_ridges() +
  facet_wrap(~attr_long) +
  theme_bw() +
  xlab(expression(delta ~ "[ % ]")) +
  ylab("") +
  scale_fill_discrete("") +
  ggtitle(expression(Plot ~ delta)) +
  theme(legend.position = "bottom")
```

```{r delta_tau}
#| message: false
#| warning: false
read_tsv("outputs/parameters.tsv") %>%
  filter(
    type == "diversity", attribute == "diversity_q1_gen",
    parameter %in% c("deltapct", "tau")
  ) %>%
  filter(!(site %in% c("Kibale", "Uppangala"))) %>%
  select(attr_long, parameter, site, median) %>%
  pivot_wider(names_from = parameter, values_from = median) %>%
  unnest(deltapct) %>%
  unnest(tau) %>%
  ggplot(aes(tau, deltapct, col = site)) +
  geom_point() +
  theme_bw() +
  ylab(expression(delta ~ "[ % ]")) +
  xlab(expression(tau ~ "[ year ]")) +
  scale_color_discrete("") +
  ggtitle(expression(Plot ~ delta ~ "~" ~ Site ~ tau)) +
  facet_wrap(~attr_long, scales = "free") +
  ggpubr::stat_cor(aes(group = NA))
```

## Observed & inferred disturbance

```{r dist_ba}
#| message: false
#| warning: false
data_rec <- read_tsv("data/derived_data/data_rec.tsv", col_types = cols()) %>%
  filter(variable == "ba") %>%
  filter(variable == "ba") %>%
  group_by(site, plot) %>%
  summarise(ba_post = min(value, na.rm = TRUE))
data_equ <- read_tsv("data/derived_data/data_equ.tsv", col_types = cols()) %>%
  filter(variable == "ba", rel_year <= 0) %>%
  group_by(site, plot) %>%
  summarise(ba_pre = median(value, na.rm = TRUE))
dist_obs <- data_equ %>%
  left_join(data_rec) %>%
  na.omit() %>%
  mutate(dist_obs = (ba_pre - ba_post) / ba_pre * 100) %>%
  select(site, plot, dist_obs) %>%
  filter(dist_obs > 0)
read_tsv("outputs/parameters.tsv") %>%
  filter(parameter == "phipct", attribute == "ba") %>%
  select(site, plot, median, sd, q5, q95) %>%
  unique() %>%
  left_join(dist_obs) %>%
  na.omit() %>%
  ggplot(aes(median, dist_obs)) +
  geom_abline(col = "darkgrey") +
  geom_segment(aes(x = median - sd, xend = median + sd)) +
  geom_segment(aes(x = q5, xend = q95), size = .1) +
  geom_point() +
  facet_wrap(~site) +
  theme_bw() +
  ggpubr::stat_cor(size = 3) +
  xlab(expression(phi[BA] ~ "inferred [ % lost ]")) +
  ylab(expression(dist[BA] ~ "observed [ % lost ]")) +
  theme(panel.spacing = unit(0, "lines"))
```

## H2 - Diversity & intensity

> **H2**: tree diversity also shows a non-linear response to increased intensity of logging: lightly logged forests show a transient increase in tree diversity, but more intensively logged forests show a strong negative effect on diversity

```{r delta_dist_ba}
#| message: false
#| warning: false
dist_p <- read_tsv("outputs/parameters.tsv") %>%
  filter(parameter == "phipct", attribute == "ba") %>%
  select(site, plot, median, sd, q5, q95) %>%
  unique() %>%
  rename_at(c("median", "sd", "q5", "q95"), ~ paste0(., "_dist_ba"))
read_tsv("outputs/parameters.tsv") %>%
  filter(!(site %in% c("Kibale", "Uppangala"))) %>%
  filter(parameter == "deltapct", type == "diversity") %>%
  left_join(dist_p) %>%
  ggplot(aes(median_dist_ba, median, col = site)) +
  geom_smooth(method = "lm", se = FALSE, col = "black", aes(group = NA)) +
  geom_point() +
  facet_wrap(~attr_long, scales = "free") +
  theme_bw() +
  xlab(expression(phi[BA] ~ "[ % lost ]")) +
  ylab(expression(delta ~ "[ % ]")) +
  scale_color_discrete(guide = "none") +
  ggpubr::stat_cor(aes(group = NA),
    label.x.npc = "left", label.y.npc = "top"
  )
```

```{r delta_dist_ba_plot}
#| message: false
#| warning: false
read_tsv("outputs/parameters.tsv") %>%
  filter(!(site %in% c("Kibale", "Uppangala"))) %>%
  filter(parameter == "deltapct", type == "diversity") %>%
  filter(attribute == "diversity_q1_gen") %>%
  left_join(dist_p) %>%
  ggplot(aes(median_dist_ba, median)) +
  geom_smooth(method = "lm", se = FALSE) +
  geom_point(col = "grey") +
  facet_wrap(~site) +
  ggpubr::stat_cor(size = 3) +
  theme_bw() +
  xlab(expression(phi[BA] ~ "[ % lost ]")) +
  ylab(expression(delta ~ "[ % ]")) +
  theme(legend.position = "bottom") +
  theme(panel.spacing = unit(0, "lines"))
```
