==> 91_stp_negative.qmd <==
```{r setup}
#| message: false
#| include: false
library(tidyverse)
```

# STP negatives {.unnumbered}

Using long term and short term processes model based on initial value $\theta_0$ we had issue with negatives values of $\mu_{rec}$ resulting in impossible log-values:

$$
\begin{align}
\mu_{rec} = \theta_{0,p} + (\theta_{\infty,s} - \theta_{0,p}) \times (LTP+STP) \\
LTP = 1 - e^{-\lambda_p \times time} \\
STP = \delta_p \times (\frac{time}{\tau_s} \times e^{1-\frac{time}{\tau_s}})^2
\end{align}
$$

Negatives values happen when the trajectory is negative and the difference of $\theta$ and the bump fall below 0, so when $time=\tau$:

$$
\begin{align}
\theta_{0,p} + (\theta_{\infty,s} - \theta_{0,p}) \times (1 - e^{-\lambda_p \times \tau_s}+\delta_p) \leq 0\\
\end{align}
$$

En $\lambda_p=0.5$ et $\tau_s=30$ , $LTP = 1-e^{-0.5*30} \approx 1$, thus:

$$
\begin{align}
\theta_{0,p} + (\theta_{\infty,s} - \theta_{0,p}) \times (1+\delta_p) \leq 0\\
\theta_{\infty,s} + (\theta_{\infty,s} - \theta_{0,p}) \times \delta_p \leq 0\\
 \delta_p \geq \frac{-\theta_{\infty,s}}  {\theta_{\infty,s} - \theta_{0,p}} ~|~ \theta_{\infty,s} - \theta_{0,p} < 0 \\
\delta_p \geq \frac{-1}  {1 - dist_p} ~|~\theta_{0,p}=dist_p\times\theta_{\infty,s} \\
\delta_p \geq \frac{1}  {dist_p-1}
\end{align}
$$

We could thus use this inequality to constrain $\delta_p$ on $dist_p$ knowing that `stan` now accepts parameters bounds to depend on another parameter. However, this is complicated as it depends on the sign in the division above, $dist=1$ is no possible, and near 1 values results in positive or negative infinite values. Thus we simply bounded $dist_p$ to 1.5 for now (50% gain of attribute post-disturbance).

```{r fig1}
#| fig-cap: "Negative STR examples with delta-dist link."
model <- function(time, dist, thetainf, lambda, tau, delta) {
  theta0 <- thetainf * dist
  ltp <- 1 - exp(-lambda * time)
  stp <- delta * (time / tau * exp(1 - time / tau))^2
  theta0 + (thetainf - theta0) * (ltp + stp)
}
data.frame(time = 1:50) %>%
  mutate(y = model(time,
    dist = 2, thetainf = 20,
    lambda = 0.5, tau = 30, delta = 0.5
  )) %>%
  mutate(y2 = model(time,
    dist = 2, thetainf = 20,
    lambda = 0.5, tau = 30, delta = 2
  )) %>%
  mutate(y3 = model(time,
    dist = 2, thetainf = 20,
    lambda = 0.5, tau = 30, delta = 1
  )) %>%
  ggplot(aes(time)) +
  geom_line(aes(y = y, col = "dist=2, delta=.5")) +
  geom_line(aes(y = y2, col = "dist=2, delta=2")) +
  geom_line(aes(y = y3, col = "dist=2, delta=1")) +
  theme_bw()
```

```{r fig2}
#| fig-cap: "delta_max with dist."
data.frame(dist = seq(.5, 2, length.out = 100)) %>%
  mutate(delta_max = 2 - (1 / (dist - 1))) %>%
  ggplot(aes(dist, delta_max)) +
  geom_line() +
  theme_bw()
```
